requirements.txt
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
beanie==1.11.1
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
fastapi-users==10.0.6
fastapi-users-db-beanie==1.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
makefun==1.13.1
motor==2.5.1
multidict==6.0.2
passlib==1.7.4
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pymongo==3.12.3
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
yarl==1.7.2
